requirements.txt
streamlit>=1.28.0
google-api-python-client>=2.104.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.1.1
transformers>=4.35.0
torch>=2.1.0
python-dotenv>=1.0.0
python-dateutil>=2.8.2
types-python-dateutil>=2.8.19